refactor(hgraph): simplify deduplicated code storage
