+++ src/algorithm/hgraph/hgraph_build.cpp
@@ -26,7 +26,6 @@
 #include "datacell/flatten_datacell_parameter.h"
 #include "dataset_impl.h"
 #include "hgraph.h"  // IWYU pragma: keep
-#include "hgraph_code_slot.h"
 #include "impl/heap/standard_heap.h"
 #include "impl/logger/logger.h"
 #include "impl/odescent/odescent_graph_builder.h"
@@ -144,7 +143,7 @@ HGraph::Build(const DatasetPtr& data) {
     this->build_cache_hit_nodes_ = 0;
     this->build_cache_missed_nodes_ = 0;
     if (this->has_loaded_cache()) {
-        if (this->using_dedup_storage()) {
+        if (this->support_duplicate_ && this->deduplicate_storage_) {
             throw VsagException(ErrorType::INVALID_ARGUMENT,
                                 "HGraph deduplicate_storage does not support build_with_cache");
         }
@@ -293,11 +292,7 @@ HGraph::Add(const DatasetPtr& data) {
     this->prepare_graph_read_codes(data, context);
     auto batch = this->prepare_add_batch(data);
     this->prepare_temporary_graph_read_codes(data, context, batch);
-    this->reserve_storage_for_batch(context, batch);
     this->insert_add_batch(data, context, batch);
-    if (context.use_dedup_storage) {
-        this->compact_physical_code_capacity_if_needed();
-    }
     return batch.failed_ids;
 }
 
@@ -314,7 +309,7 @@ HGraph::validate_add_data(const DatasetPtr& data) const {
 HGraph::AddContext
 HGraph::prepare_add_context(const DatasetPtr& data) {
     AddContext context;
-    context.use_dedup_storage = this->using_dedup_storage();
+    context.use_dedup_storage = this->support_duplicate_ && this->deduplicate_storage_;
     context.need_temporary_sq8_build_data =
         need_temporary_sq8_build_data(this->basic_flatten_codes_, this->has_precise_reorder());
     context.use_parallel_add = this->thread_pool_ != nullptr;
@@ -429,15 +424,6 @@ HGraph::prepare_temporary_graph_read_codes(const DatasetPtr& data,
 }
 
 void
-HGraph::reserve_storage_for_batch(const AddContext& context, const AddBatch& batch) {
-    if (context.use_dedup_storage and not batch.rows.empty()) {
-        auto required_capacity =
-            this->code_slot_map_->PhysicalCount() + static_cast<InnerIdType>(batch.rows.size());
-        this->ensure_physical_code_capacity(required_capacity);
-    }
-}
-
-void
 HGraph::insert_add_batch(const DatasetPtr& data, const AddContext& context, const AddBatch& batch) {
     std::vector<std::future<void>> futures;
     const auto* extra_infos = data->GetExtraInfos();
@@ -506,12 +492,12 @@ HGraph::insert_persistent_codes_unlocked(const void* data, InnerIdType inner_id)
 
 void
 HGraph::insert_persistent_codes_to_slot(const void* data, InnerIdType code_slot_id) {
-    InsertVectorToHGraphPhysicalSlot(this->basic_flatten_codes_, data, code_slot_id);
+    InsertVectorToHGraphCodeSlot(this->basic_flatten_codes_, data, code_slot_id);
     if (has_precise_reorder()) {
-        InsertVectorToHGraphPhysicalSlot(this->high_precise_codes_, data, code_slot_id);
+        InsertVectorToHGraphCodeSlot(this->high_precise_codes_, data, code_slot_id);
     }
     if (create_new_raw_vector_) {
-        InsertVectorToHGraphPhysicalSlot(this->raw_vector_, data, code_slot_id);
+        InsertVectorToHGraphCodeSlot(this->raw_vector_, data, code_slot_id);
     }
 }
 
@@ -519,9 +505,7 @@ bool
 HGraph::insert_one_logical_point(const void* data, const AddRow& row, const AddContext& context) {
     const auto inner_id = row.inner_id;
     const auto level = row.level;
-    if (not context.use_dedup_storage) {
-        this->insert_persistent_codes(data, inner_id);
-    }
+    this->prepare_codes_before_probe_if_needed(data, inner_id, context);
 
     auto make_search_param = [&]() {
         InnerSearchParam param;
@@ -553,9 +537,28 @@ HGraph::insert_one_logical_point(const void* data, const AddRow& row, const AddC
         return true;
     }
 
-    this->publish_unique_storage_if_needed(data, inner_id, context, rlock);
-    this->publish_unique_to_graphs(data, level, inner_id, param, probe, context);
+    this->publish_unique_under_shared_global_lock(
+        data, level, inner_id, param, probe, context, rlock);
     return true;
+}
+
+void
+HGraph::prepare_codes_before_probe_if_needed(const void* data,
+                                             InnerIdType inner_id,
+                                             const AddContext& context) {
+    if (not context.use_dedup_storage) {
+        this->insert_persistent_codes(data, inner_id);
+    }
+}
+
+void
+HGraph::publish_duplicate_storage_if_needed(InnerIdType group_id,
+                                            InnerIdType duplicate_id,
+                                            const AddContext& context) {
+    if (context.use_dedup_storage) {
+        auto code_slot_id = this->code_slot_map_->Resolve(group_id);
+        this->code_slot_map_->PublishSlot(duplicate_id, code_slot_id);
+    }
 }
 
 bool
@@ -567,12 +570,8 @@ HGraph::publish_duplicate_if_found(const GraphAddProbeResult& probe,
     }
     auto group_id = static_cast<InnerIdType>(probe.duplicate_id);
     auto duplicate_id = inner_id;
-    if (context.use_dedup_storage) {
-        auto code_slot_id = this->code_slot_map_->Resolve(group_id);
-        this->code_slot_map_->PublishSlot(duplicate_id, code_slot_id);
-    }
-    std::unique_lock lock(this->label_lookup_mutex_);
-    bottom_graph_->SetDuplicateId(group_id, duplicate_id);
+    this->publish_duplicate_storage_if_needed(group_id, duplicate_id, context);
+    this->publish_duplicate_to_tracker(group_id, duplicate_id);
     return true;
 }
 
@@ -621,6 +620,18 @@ HGraph::ensure_route_graphs_for_level(int level) {
 }
 
 void
+HGraph::publish_unique_under_shared_global_lock(const void* data,
+                                                int level,
+                                                InnerIdType inner_id,
+                                                InnerSearchParam& param,
+                                                const GraphAddProbeResult& probe,
+                                                const AddContext& context,
+                                                std::shared_lock<std::shared_mutex>& read_lock) {
+    this->publish_unique_storage_if_needed(data, inner_id, context, read_lock);
+    this->publish_unique_to_graphs(data, level, inner_id, param, probe, context);
+}
+
+void
 HGraph::publish_unique_under_unique_global_lock(const void* data,
                                                 int level,
                                                 InnerIdType inner_id,
@@ -665,8 +676,9 @@ HGraph::probe_graph_for_add(const void* data,
     param.topk = static_cast<int64_t>(ef_construct_);
     if (this->support_duplicate_) {
         param.find_duplicate = true;
-        param.duplicate_query_id =
-            this->using_dedup_storage() ? std::numeric_limits<InnerIdType>::max() : inner_id;
+        param.duplicate_query_id = this->support_duplicate_ && this->deduplicate_storage_
+                                       ? std::numeric_limits<InnerIdType>::max()
+                                       : inner_id;
         param.duplicate_distance_threshold = this->duplicate_distance_threshold_;
     }
 
@@ -753,7 +765,7 @@ HGraph::publish_unique_to_route_graphs(const void* data,
 
 void
 HGraph::ensure_physical_code_capacity(InnerIdType required_capacity) {
-    if (not this->using_dedup_storage()) {
+    if (not(this->support_duplicate_ && this->deduplicate_storage_)) {
         return;
     }
     if (this->physical_code_capacity_.load(std::memory_order_acquire) >= required_capacity) {
@@ -765,14 +777,15 @@ HGraph::ensure_physical_code_capacity(InnerIdType required_capacity) {
 
 void
 HGraph::ensure_physical_code_capacity_unlocked(InnerIdType required_capacity) {
-    if (not this->using_dedup_storage()) {
+    if (not(this->support_duplicate_ && this->deduplicate_storage_)) {
         return;
     }
     if (this->physical_code_capacity_.load(std::memory_order_acquire) >= required_capacity) {
         return;
     }
 
-    auto new_capacity = this->align_physical_code_capacity(required_capacity);
+    auto new_capacity = static_cast<InnerIdType>(
+        next_multiple_of_power_of_two(required_capacity, this->resize_increase_count_bit_));
     GetHGraphPhysicalFlatten(this->basic_flatten_codes_)->Resize(new_capacity);
     if (has_precise_reorder()) {
         GetHGraphPhysicalFlatten(this->high_precise_codes_)->Resize(new_capacity);
@@ -784,51 +797,10 @@ HGraph::ensure_physical_code_capacity_unlocked(InnerIdType required_capacity) {
     this->cal_memory_usage();
 }
 
-InnerIdType
-HGraph::align_physical_code_capacity(InnerIdType required_capacity) const {
-    if (required_capacity == 0) {
-        return 0;
-    }
-    auto block_size = Options::Instance().block_size_limit();
-    auto code_size =
-        std::max<uint64_t>(1, static_cast<uint64_t>(this->basic_flatten_codes_->code_size_));
-    auto slots_per_block = std::max<uint64_t>(1, block_size / code_size);
-    auto required = static_cast<uint64_t>(required_capacity);
-    return static_cast<InnerIdType>(((required + slots_per_block - 1) / slots_per_block) *
-                                    slots_per_block);
-}
-
 void
-HGraph::compact_physical_code_capacity_if_needed() {
-    if (not this->using_dedup_storage()) {
-        return;
-    }
-
-    auto block_slots = this->align_physical_code_capacity(1);
-    auto physical_count = this->code_slot_map_->PhysicalCount();
-    auto target_capacity = this->align_physical_code_capacity(physical_count);
-    auto current_capacity = this->physical_code_capacity_.load(std::memory_order_acquire);
-    if (current_capacity <= target_capacity || current_capacity - target_capacity < block_slots) {
-        return;
-    }
-
-    std::scoped_lock lock(this->global_mutex_);
-    physical_count = this->code_slot_map_->PhysicalCount();
-    target_capacity = this->align_physical_code_capacity(physical_count);
-    current_capacity = this->physical_code_capacity_.load(std::memory_order_acquire);
-    if (current_capacity <= target_capacity || current_capacity - target_capacity < block_slots) {
-        return;
-    }
-
-    GetHGraphPhysicalFlatten(this->basic_flatten_codes_)->ShrinkToFit(target_capacity);
-    if (has_precise_reorder()) {
-        GetHGraphPhysicalFlatten(this->high_precise_codes_)->ShrinkToFit(target_capacity);
-    }
-    if (create_new_raw_vector_) {
-        GetHGraphPhysicalFlatten(this->raw_vector_)->ShrinkToFit(target_capacity);
-    }
-    this->physical_code_capacity_.store(target_capacity, std::memory_order_release);
-    this->cal_memory_usage();
+HGraph::publish_duplicate_to_tracker(InnerIdType group_id, InnerIdType duplicate_id) {
+    std::unique_lock lock(this->label_lookup_mutex_);
+    bottom_graph_->SetDuplicateId(group_id, duplicate_id);
 }
 
 void
@@ -846,10 +818,10 @@ HGraph::resize(uint64_t new_size) {
         pool_ = std::make_shared<VisitedListPool>(1, allocator_, new_size_power_2, allocator_);
         this->label_table_->Resize(new_size_power_2);
         bottom_graph_->Resize(new_size_power_2);
-        if (this->using_dedup_storage()) {
+        if (this->support_duplicate_ && this->deduplicate_storage_) {
             this->code_slot_map_->ReserveLogicalSize(static_cast<InnerIdType>(new_size_power_2));
         }
-        if (not this->using_dedup_storage()) {
+        if (not(this->support_duplicate_ && this->deduplicate_storage_)) {
             this->basic_flatten_codes_->Resize(new_size_power_2);
             if (has_precise_reorder()) {
                 this->high_precise_codes_->Resize(new_size_power_2);
