+++ src/algorithm/hgraph/hgraph_parameter.cpp
@@ -142,6 +142,12 @@ HGraphParameter::FromJson(const JsonType& json) {
     if (json.Contains(SUPPORT_FORCE_REMOVE)) {
         this->support_force_remove = json[SUPPORT_FORCE_REMOVE].GetBool();
     }
+    if (this->deduplicate_storage && this->support_force_remove) {
+        throw VsagException(
+            ErrorType::INVALID_ARGUMENT,
+            "deduplicate_storage does not support force remove because duplicate groups share "
+            "physical vector slots");
+    }
     if (json.Contains(HGRAPH_PERSIST_SOURCE_ID_KEY)) {
         this->persist_source_id = json[HGRAPH_PERSIST_SOURCE_ID_KEY].GetBool();
     }
