+++ src/algorithm/hgraph/hgraph_serialize.cpp
@@ -21,7 +21,6 @@
 #include "common.h"
 #include "datacell/sparse_graph_datacell.h"
 #include "hgraph.h"  // IWYU pragma: keep
-#include "hgraph_code_slot.h"
 #include "impl/heap/standard_heap.h"
 #include "impl/odescent/odescent_graph_builder.h"
 #include "impl/pruning_strategy.h"
@@ -359,7 +358,7 @@ HGraph::Serialize(StreamWriter& writer) const {
 
     // FIXME(wxyu): this option is used for special purposes, like compatibility testing
     if (this->use_old_serial_format_) {
-        if (this->using_dedup_storage()) {
+        if (this->support_duplicate_ && this->deduplicate_storage_) {
             throw VsagException(ErrorType::INVALID_ARGUMENT,
                                 "HGraph duplicate code slot mapping does not support v0.14 "
                                 "serialization");
@@ -383,7 +382,7 @@ HGraph::Serialize(StreamWriter& writer) const {
     }
 
     this->serialize_label_info(writer);
-    if (this->using_dedup_storage()) {
+    if (this->support_duplicate_ && this->deduplicate_storage_) {
         this->code_slot_map_->Serialize(writer);
     }
     this->basic_flatten_codes_->Serialize(writer);
@@ -782,7 +781,7 @@ HGraph::Deserialize(StreamReader& reader) {
         logger::debug("parse with v0.14 version format");
 
         this->deserialize_basic_info_v0_14(reader);
-        if (this->using_dedup_storage()) {
+        if (this->support_duplicate_ && this->deduplicate_storage_) {
             throw VsagException(ErrorType::INVALID_ARGUMENT,
                                 "HGraph duplicate code slot mapping does not support v0.14 "
                                 "serialization");
@@ -803,7 +802,7 @@ HGraph::Deserialize(StreamReader& reader) {
         }
         auto new_size = max_capacity_.load();
         this->neighbors_mutex_->Resize(new_size);
-        if (this->using_dedup_storage()) {
+        if (this->support_duplicate_ && this->deduplicate_storage_) {
             this->code_slot_map_->ReserveLogicalSize(static_cast<InnerIdType>(new_size));
         }
 
@@ -841,7 +840,7 @@ HGraph::Deserialize(StreamReader& reader) {
         this->label_table_->is_legacy_duplicate_format_ = (dup_version == 0);
 
         this->deserialize_label_info(buffer_reader);
-        if (this->using_dedup_storage()) {
+        if (this->support_duplicate_ && this->deduplicate_storage_) {
             this->code_slot_map_->Deserialize(buffer_reader);
             this->total_count_.store(this->code_slot_map_->PublishedLogicalCount(),
                                      std::memory_order_release);
@@ -862,7 +861,7 @@ HGraph::Deserialize(StreamReader& reader) {
         }
         auto new_size = max_capacity_.load();
         this->neighbors_mutex_->Resize(new_size);
-        if (this->using_dedup_storage()) {
+        if (this->support_duplicate_ && this->deduplicate_storage_) {
             this->code_slot_map_->ReserveLogicalSize(static_cast<InnerIdType>(new_size));
         }
 
@@ -871,7 +870,7 @@ HGraph::Deserialize(StreamReader& reader) {
         if (this->extra_info_size_ > 0 && this->extra_infos_ != nullptr) {
             this->extra_infos_->Deserialize(buffer_reader);
         }
-        if (not this->using_dedup_storage()) {
+        if (not(this->support_duplicate_ && this->deduplicate_storage_)) {
             this->total_count_ = this->basic_flatten_codes_->TotalCount();
         }
 
@@ -884,6 +883,36 @@ HGraph::Deserialize(StreamReader& reader) {
         }
         if (this->raw_vector_ != nullptr) {
             this->has_raw_vector_ = true;
+        }
+    }
+    if (this->using_dedup_storage()) {
+        auto logical_count = this->code_slot_map_->PublishedLogicalCount();
+        auto physical_count = this->code_slot_map_->PhysicalCount();
+        auto validate_physical_count = [physical_count](const FlattenInterfacePtr& flatten,
+                                                        const char* name) {
+            auto actual_count = GetHGraphPhysicalFlatten(flatten)->TotalCount();
+            if (actual_count != physical_count) {
+                throw VsagException(
+                    ErrorType::INVALID_BINARY,
+                    fmt::format("deduplicated HGraph {} physical count mismatch: {} != {}",
+                                name,
+                                actual_count,
+                                physical_count));
+            }
+        };
+        if (logical_count != this->total_count_.load(std::memory_order_acquire)) {
+            throw VsagException(ErrorType::INVALID_BINARY,
+                                "deduplicated HGraph code-slot logical count is inconsistent");
+        }
+        validate_physical_count(this->basic_flatten_codes_, "base codes");
+        if (this->has_precise_reorder()) {
+            validate_physical_count(this->high_precise_codes_, "precise codes");
+        }
+        if (this->create_new_raw_vector_) {
+            validate_physical_count(this->raw_vector_, "raw vectors");
+        }
+        for (InnerIdType inner_id = 0; inner_id < logical_count; ++inner_id) {
+            (void)this->code_slot_map_->Resolve(inner_id);
         }
     }
     this->cal_memory_usage();
